% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extract.R
\name{Extract}
\alias{Extract}
\title{Extracting specific columns from a data\cr}
\usage{
Extract(X, varind, interaction.ind)
}
\arguments{
\item{interaction.ind}{}
}
\description{
Extracting specific columns from a data\cr
}
