% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial.R
\name{initial}
\alias{initial}
\title{Setting up initial candidate models\cr}
\usage{
initial(
  X,
  y,
  heredity = "Strong",
  nmain.p,
  r1,
  r2,
  interaction.ind = NULL,
  pi1 = 0.32,
  pi2 = 0.32,
  pi3 = 0.32,
  lambda = 10,
  q = 40
)
}
\arguments{
\item{X}{Input data. An optional data frame, or numeric matrix of dimension
\code{n} by \code{nmain.p}. Note that the two-way interaction effects should not
be included in \code{X} because this function automatically generates the
corresponding two-way interaction effects if needed.}

\item{y}{Response variable. A \code{n}-dimensional vector, where \code{n} is the number
of observations in \code{X}.}

\item{heredity}{whether to enforce Strong, Weak, or No heredity.
Default is "Strong".}

\item{nmain.p}{A numeric value that represents the total number of main effects
in \code{X}.}

\item{r1}{A numeric value indicating the maximum number of main effects.}

\item{r2}{A numeric value indicating the maximum number of interaction effects.}

\item{interaction.ind}{A two-column numeric matrix containing all possible
two-way interaction effects. It must be generated outside of this function
using \code{t(utils::combn())}. See Example section for details.}

\item{pi1}{A numeric value between 0 and 1, defined by users. Default is 0.32.
For guidance on selecting an appropriate value, please refer to the Details section.}

\item{pi2}{A numeric value between 0 and 1, defined by users. Default is 0.32.
For guidance on selecting an appropriate value, please refer to the Details section.}

\item{pi3}{A numeric value between 0 and 1, defined by users. Default is 0.32.
For guidance on selecting an appropriate value, please refer to the Details section.}

\item{lambda}{A numeric value defined by users. Default is 10.
For guidance on selecting an appropriate value, please refer to the Details section.}

\item{q}{A numeric value indicating the number of models in each generation (e.g.,
the population size). Default is 40.}
}
\value{
A \code{list} of output. The components are:
\item{initialize}{Initial candidate models. A numeric matrix of dimension \code{q} by
\code{r1+r2} where each row represents a fitted model. Duplicated models are allowed.}
\item{InterRank}{Rank of all candidate interaction effects. A two-column numeric
matrix. The first column contains indices of ranked two-way interaction effects, and the
second column contains its corresponding ABC score.}
\item{mainind.sel}{Selected main effects.  A \code{r1}-dimensional vector.}
\item{mainpool}{Ranked main effects in \code{X}.}
}
\description{
This function automatically ranks all candidate interaction effects under
Strong, Weak, or No heredity condition and obtains initial candidate models.
}
\details{
\itemize{
\item For inputs \code{pi1}, \code{pi2}, and \code{pi3}, the number needs to
satisfy the condition: \eqn{\pi_1+\pi_2+\pi_3=1-\pi_0} where \eqn{\pi_0}
is a numeric value between 0 and 1, the smaller the better.
\item For input \code{lambda}, the number needs to satisfy the condition:
\eqn{\lambda\geq 5.1/log(2)}.
\item For input \code{q}, this number is fixed throughout each generation.
}
}
\examples{
# Under Strong heredity
set.seed(0)
nmain.p <- 4
interaction.ind <- t(combn(4,2))
X <- matrix(rnorm(50*4,1,0.1), 50, 4)
epl <- rnorm(50,0,0.01)
y<- 1+X[,1]+X[,2]+X[,1]*X[,2]+epl
p1 <- initial(X, y, nmain.p = 4, r1 = 3, r2 = 3,
    interaction.ind = interaction.ind, q = 5)
p1
}
\seealso{
\code{\link{ABC}}, \code{\link{Extract}}.
}
